refactor: consolidate picking file into frontend
